Use POST /subscriptions/id/cancel to cancel subscription
Up until now, we used a DELETE to /subscriptions/id to cancel a
subscription. Recently, we added a password check to this endpoint,
which is passed through the body. It is however not clear whether this
is valid according to the HTTP spec, and not all clients seem to
support this. Because of this, we chose to change the endpoint to a
POST on /subscriptions/id/cancel instead.