Additional status and type for order
An order can now be of the type: silent_imsi_sim_swap
which is a sim_swap of a special silent imsi card. This needs to
follow a different flow to enable so a distinct type is need and
sim_swap cannot be reused

An order can now be in the state: waiting_for_sms
This means a special silent imsi card has been send and it needs to
activate via sms a regular card can only be activated via the website.